# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LTT Sales Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales-logs:
    get:
      operationId: listSalesLogs
      tags: [salesLogs]
      summary: List inspection sales logs
      parameters:
        - in: query
          name: agentId
          schema:
            type: integer
        - in: query
          name: inspectorId
          schema:
            type: integer
        - in: query
          name: complianceFlag
          schema:
            type: string
      responses:
        "200":
          description: List of sales logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesLog"
    post:
      operationId: createSalesLog
      tags: [salesLogs]
      summary: Create a sales log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesLogBody"
      responses:
        "201":
          description: Created sales log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesLog"
  /sales-logs/{id}:
    get:
      operationId: getSalesLog
      tags: [salesLogs]
      summary: Get sales log by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sales log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesLog"
        "404":
          description: Not found
    patch:
      operationId: updateSalesLog
      tags: [salesLogs]
      summary: Update sales log
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalesLogBody"
      responses:
        "200":
          description: Updated sales log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesLog"

  # TICKETS
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: assignedTo
          schema:
            type: integer
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketBody"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Not found
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketBody"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  # INVENTORY
...
  /scores/{agentId}:
    get:
      operationId: getAgentScore
      tags: [scores]
      summary: Get score for a specific agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentScore"
        "404":
          description: Not found
  /scores/{agentId}/recalculate:
    post:
      operationId: recalculateAgentScore
      tags: [scores]
      summary: Recalculate score for an agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recalculated score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentScore"

  # NOTIFICATIONS
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    SalesLog:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        inspectorId:
          type: integer
        visitId:
          type: ["integer", "null"]
        date:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        rechargeCardsQty:
          type: ["integer", "null"]
        rechargeCardsValue:
          type: ["number", "null"]
        devicesQty:
          type: ["integer", "null"]
        simCardsQty:
          type: ["integer", "null"]
        ftthCount:
          type: ["integer", "null"]
        adslCount:
          type: ["integer", "null"]
        observedDailySalesValue:
          type: number
        reportedSalesValue:
          type: number
        variance:
          type: number
        notes:
          type: ["string", "null"]
        complianceFlag:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - inspectorId
        - date
        - observedDailySalesValue
        - reportedSalesValue
        - variance
        - complianceFlag
        - createdAt
    CreateSalesLogBody:
      type: object
      properties:
        agentId:
          type: integer
        inspectorId:
          type: integer
        visitId:
          type: ["integer", "null"]
        date:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        rechargeCardsQty:
          type: ["integer", "null"]
        rechargeCardsValue:
          type: ["number", "null"]
        devicesQty:
          type: ["integer", "null"]
        simCardsQty:
          type: ["integer", "null"]
        ftthCount:
          type: ["integer", "null"]
        adslCount:
          type: ["integer", "null"]
        observedDailySalesValue:
          type: number
        reportedSalesValue:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - agentId
        - inspectorId
        - date
        - observedDailySalesValue
        - reportedSalesValue
    UpdateSalesLogBody:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        complianceFlag:
          type: ["string", "null"]
        observedDailySalesValue:
          type: ["number", "null"]
        reportedSalesValue:
          type: ["number", "null"]

    # TICKETS
    Ticket:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        category:
          type: string
        createdById:
          type: integer
        assignedToId:
          type: ["integer", "null"]
        agentId:
          type: ["integer", "null"]
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - status
        - priority
        - category
        - createdById
        - createdAt
        - updatedAt
    CreateTicketBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        category:
          type: string
        createdById:
          type: integer
        assignedToId:
          type: ["integer", "null"]
        agentId:
          type: ["integer", "null"]
      required:
        - title
        - description
        - priority
        - category
        - createdById
    UpdateTicketBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        resolvedAt:
          type: ["string", "null"]
          format: date-time

    # INVENTORY
...
    AgentScore:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        agentName:
          type: string
        complianceScore:
          type: number
        salesAccuracyScore:
          type: number
        salesPerformanceScore:
          type: number
        activityScore:
          type: number
        finalScore:
          type: number
        riskLevel:
          type: string
        classification:
          type: string
        recommendation:
          type: string
        lastUpdated:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - agentName
        - complianceScore
        - salesAccuracyScore
        - salesPerformanceScore
        - activityScore
        - finalScore
        - riskLevel
        - classification
        - recommendation
        - lastUpdated

    # NOTIFICATIONS
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - message
        - isRead
        - createdAt

    # DASHBOARD
...